refactor: move file out of helper
